Minor update to calcTruckloadEquivalencies()
Under certain conditions the final calculation of loaded trucks table in calcTruckloadEquivalencies would create the value field as integer64, whereas in empty table it is always numeric. I've added as explicit cast to numeric for the former so that it doesn't trip up rbindlist(), which doParallel does a poor job of handling when it occurs.